==> Euro2024/1.Groups_2024.ipynb <==
!pip install pandas
from string import ascii_uppercase as alfabeto
import pickle
import pandas as pd

# read_html allows to scrape data
tablas = pd.read_html('https://web.archive.org/web/20240605004535/https://en.wikipedia.org/wiki/UEFA_Euro_2024')

# Find all the groups
tablas[19]
tablas[26]
tablas[33]
tablas[40]
tablas[47]
tablas[54]

tablas[19]

# Get all the groups in the tournament
# Rename column Puntos as Pts
# Create a dictionary to insert the groups
dict_tablas = {}
for letra, i in zip(alfabeto, range(19, 60, 7)):
  df = tablas[i]
  df.rename(columns={df.columns[1]: 'Team'}, inplace=True)
  dict_tablas[f'Group {letra}'] = df
  print(df,'\n')

tablas[19]

with open('dict_table', 'wb') as output:
  pickle.dump(dict_tablas, output)

dict_tablas
