==> Euro2024/2.European_Championship_1960-2020.ipynb <==
from bs4 import BeautifulSoup
import requests
import pandas as pd

years = [1960, 1964, 1968, 1972,1976, 1980, 1984, 1988,
         1992, 1996, 2000, 2004, 2008, 2012, 2016, 2020]

# Método para obtener todos los partidos de todas las ediciones de Eurocopa jugadas
def get_matches(year):
  web = f'https://en.wikipedia.org/wiki/UEFA_Euro_{year}'
  response = requests.get(web)
  content = response.text

  soup = BeautifulSoup(content, 'lxml')
  # En el html sacamos un patrón que se repite para todos los partidos de futbol
  matches = soup.find_all('div', class_='footballbox')

  # Creamos tres listas para almacenar los datos
  home = []
  score = []
  away = []

  # matches es una lista y queremos sacar el nombre del equipo local, el resultado y el equipo visitante
  for match in matches:
    home.append(match.find('th', class_='fhome').get_text())
    score.append(match.find('th', class_='fscore').get_text())
    away.append(match.find('th', class_='faway').get_text())

  dict_football = {'home': home, 'score': score, 'away': away}

  df_football = pd.DataFrame(dict_football)
  # Añadimos la columna year
  df_football['year'] = year

  return df_football

uefa = [get_matches(year) for year in years]

#Exportamos los datos de todos los mundiales a un fichero
df_uefa = pd.concat(uefa, ignore_index=True)
df_uefa.to_csv('uefa_europeancup_historical_data.csv', index=False)

df_fixture = get_matches('2024')
df_fixture.to_csv('uefa_europeancup_fixture.csv', index=False)
