==> Euro2024/3.Clean_Data.ipynb <==
import pandas as pd

df_historical_data = pd.read_csv('uefa_europeancup_historical_data.csv')
df_fixture = pd.read_csv('uefa_europeancup_fixture.csv')


# strip deletes blank spaces at the beginning and at the end
df_fixture['home'] = df_fixture['home'].str.strip()
df_fixture['away'] = df_fixture['away'].str.strip()

df_fixture[df_fixture['home'].isnull()]
# There is no null value so we will not delete any fixture

# Delete unwanted text from the score column (e.g. a.e.t, or Italy's win on coin toss)
df_historical_data['score'] = df_historical_data['score'].str.replace('[^\d–]', '', regex=True)

df_historical_data['home'] = df_historical_data['home'].str.strip()
df_historical_data['away'] = df_historical_data['away'].str.strip()

# Separate score column into HomeGoals and AwayGoals
df_historical_data[['HomeGoals', 'AwayGoals']] = df_historical_data['score'].str.split('–', expand=True)

# Remove score column
df_historical_data.drop('score', axis=1, inplace=True)

# inplace=True to update the dataframe. We could also assign it to the df_historical_data variable
df_historical_data.rename(columns={'home': 'HomeTeam', 'away': 'AwayTeam', 'year': 'Year'}, inplace=True)

# Convert HomeGoals to integer values (they were object type)
df_historical_data = df_historical_data.astype({'HomeGoals': int, 'AwayGoals': int, 'Year': int})

df_historical_data.dtypes

# New column that adds home and away goals
df_historical_data['TotalGoals'] = df_historical_data['HomeGoals'] + df_historical_data['AwayGoals']

# Exporting clean dataframes
df_historical_data.to_csv('clean_uefa_europeancup_historical_data.csv', index=False)
df_fixture.to_csv('clean_uefa_europeancup_fixture.csv', index=False)

# Extra verifications to check the number of matches played in each edition
years = [1960, 1964, 1968, 1972,1976, 1980, 1984, 1988,
         1992, 1996, 2000, 2004, 2008, 2012, 2016, 2020]

for year in years:
  print(year, len(df_historical_data[df_historical_data['Year'] == year]))
