==> Euro2024/4.Prediction.ipynb <==
import pandas as pd
import pickle
from scipy.stats import poisson

# Load the files previously cleaned and organised
dict_table = pickle.load(open('dict_table', 'rb'))
df_historical_data = pd.read_csv('clean_uefa_europeancup_historical_data.csv')
df_fixture = pd.read_csv('clean_uefa_europeancup_fixture.csv')

df_home = df_historical_data[['HomeTeam', 'HomeGoals', 'AwayGoals']]
df_away = df_historical_data[['AwayTeam', 'HomeGoals', 'AwayGoals']]

# Rename columns
# The away team will have its goals scored in AwayGoals
df_home = df_home.rename(columns={'HomeTeam': 'Team', 'HomeGoals': 'GoalsScored', 'AwayGoals': 'GoalsConceded'})
df_away = df_away.rename(columns={'AwayTeam': 'Team', 'HomeGoals': 'GoalsConceded', 'AwayGoals': 'GoalsScored'})

dict_table

# Metric used to differentiate each team (CIS is a region formed after the dissolution of the URSS that participated in 1992)
df_team_strength = pd.concat([df_home, df_away], ignore_index=True).groupby('Team').mean()

df_team_strength

# Method to predict the points each team will get
# lambda will be the median of goals in 90 minutes (one lambda for each team)
def predict_points(home, away):
  if home in df_team_strength.index and away in df_team_strength.index:
    # goals_scored * goals_conceded
    lamb_home = df_team_strength.at[home, 'GoalsScored'] * df_team_strength.at[away, 'GoalsConceded']
    lamb_away = df_team_strength.at[away, 'GoalsScored'] * df_team_strength.at[home, 'GoalsConceded']

    prob_home, prob_away, prob_draw = 0, 0, 0
    # Let's say a team can only score up to 10 goals
    # x is the number of goals scored by the home team
    # y is the number of goals scored by the away team
    for x in range(0, 11):
      for y in range(0, 11):
        p = poisson.pmf(x, lamb_home) * poisson.pmf(y, lamb_away)
        if x == y:
          prob_draw += p
        elif x > y:
          prob_home += p
        else:
          prob_away += p

    points_home = 3 * prob_home + prob_draw
    points_away = 3 * prob_away + prob_draw
    return (points_home, points_away)

  else:
    return (0,0)


# Testing function
# Spain and Italy are in the same group
# predict_points('Spain', 'Italy')
# Portugal and Georgia are in the same group but the latest have never participated before so they do not appear in df_team_strength
# predict_points('Portugal', 'Georgia') Result: (0, 0)
# Serbia has never participated as an independent country too

df_fixture_group_stage = df_fixture[:36].copy()
df_fixture_round_of_16 = df_fixture[36:44].copy()
df_fixture_quarter_finals = df_fixture[44:48].copy()
df_fixture_semi_finals = df_fixture[48:50].copy()
df_fixture_final = df_fixture[50:].copy()

for group in dict_table:
  teams_in_group = dict_table[group]['Team'].values
  # Get the six matches every group has to avoid unnecessary simulations
  df_fixture_six = df_fixture_group_stage[df_fixture_group_stage['home'].isin(teams_in_group)]
  for index, row in df_fixture_six.iterrows():
    home, away = row['home'], row['away']
    points_home, points_away = predict_points(home, away)
    dict_table[group].loc[dict_table[group]['Team'] == home, 'Pts'] += points_home
    dict_table[group].loc[dict_table[group]['Team'] == away, 'Pts'] += points_away

  # The team with the highest score will be at the top of the table
  dict_table[group] = dict_table[group].sort_values('Pts', ascending=False).reset_index()
  dict_table[group] = dict_table[group][['Team', 'Pts']]
  dict_table[group] = dict_table[group].round(0)

df_fixture_round_of_16

for group in dict_table:
  group_winner = dict_table[group].loc[0, 'Team']
  runner_up = dict_table[group].loc[1, 'Team']

  # Replace group winner and runner-up in the round of 16 fixtures
  df_fixture_round_of_16.replace({f'Winner {group}': group_winner, f'Runner-up {group}': runner_up}, inplace=True)
  # Replace manually the best third-place teams
  df_fixture_round_of_16.replace({f'3rd Group D/E/F': 'Poland'}, inplace=True)
  df_fixture_round_of_16.replace({f'3rd Group A/D/E/F': 'Ukraine'}, inplace=True)
  df_fixture_round_of_16.replace({f'3rd Group A/B/C': 'Slovenia'}, inplace=True)
  df_fixture_round_of_16.replace({f'3rd Group A/B/C/D': 'Croatia'}, inplace=True)

# Add winner column
df_fixture_round_of_16['winner'] = '?'
df_fixture_round_of_16

def get_winner(df_fixture_updated):
  for index, row in df_fixture_updated.iterrows():
    home, away = row['home'], row['away']
    points_home, points_away = predict_points(home, away)
    if points_home > points_away:
      winner = home
    else:
      winner = away
    df_fixture_updated.loc[index, 'winner'] = winner

  return df_fixture_updated

get_winner(df_fixture_round_of_16)

def update_table(df_fixture_octavos, df_fixture_cuartos):
  for index, row in df_fixture_octavos.iterrows():
    winner = df_fixture_octavos.loc[index, 'winner']
    match = df_fixture_octavos.loc[index, 'score']
    df_fixture_cuartos.replace({f'Winner {match}': winner}, inplace=True)
  df_fixture_cuartos['winner'] = '?'

  return df_fixture_cuartos

update_table(df_fixture_round_of_16, df_fixture_quarter_finals)

get_winner(df_fixture_quarter_finals)

update_table(df_fixture_quarter_finals, df_fixture_semi_finals)

get_winner(df_fixture_semi_finals)

update_table(df_fixture_semi_finals, df_fixture_final)

get_winner(df_fixture_final)
